+++ src/commands/generate.rs
@@ -1,116 +1,57 @@
-//! Implementation of the `generate` command.
+//! Implementation of the `docset` subcommand.
 
-use crate::{
-    common::*,
-    error::*
-};
+use crate::{error::*, io::*, DocsetParams};
 
+use cargo_metadata::Metadata;
+use derive_more::Constructor;
 use rusqlite::{params, Connection};
-use snafu::ResultExt;
-use toml::Value;
+use snafu::{ensure, ResultExt};
 
 use std::{
     borrow::ToOwned,
     ffi::OsStr,
+    fmt::Display,
     fs::{copy, create_dir_all, read_dir, remove_dir_all, File},
-    io::{Read, Write},
+    io::Write,
     path::{Path, PathBuf},
     process::Command,
-    str::FromStr
+    result::Result as StdResult,
 };
 
-#[derive(Debug)]
-pub struct GenerateConfig {
-    pub manifest_path: Option<String>,
-    pub package: Package,
-    pub no_dependencies: bool,
-    pub doc_private_items: bool,
-    pub features: Vec<String>,
-    pub no_default_features: bool,
-    pub all_features: bool,
-    pub target: Option<String>,
-    pub target_dir: Option<String>,
-    pub exclude: Vec<String>,
-    pub clean: bool,
-    pub lib: bool,
-    pub bin: Vec<String>,
-    pub bins: bool
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum EntryType {
+    Constant,
+    Enum,
+    Function,
+    Macro,
+    Module,
+    Package, // i.e. crate
+    Struct,
+    Trait,
+    Type //Union // Is this even implemented in Rust ?
 }
 
-impl Default for GenerateConfig {
-    fn default() -> GenerateConfig {
-        GenerateConfig {
-            manifest_path: None,
-            package: Package::Current,
-            no_dependencies: false,
-            doc_private_items: false,
-            exclude: Vec::new(),
-            features: Vec::new(),
-            no_default_features: false,
-            all_features: false,
-            target: None,
-            target_dir: None,
-            clean: true,
-            lib: false,
-            bin: Vec::new(),
-            bins: false
+impl Display for EntryType {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> StdResult<(), std::fmt::Error> {
+        match self {
+            EntryType::Constant => write!(f, "Constant"),
+            EntryType::Enum => write!(f, "Enum"),
+            EntryType::Function => write!(f, "Function"),
+            EntryType::Macro => write!(f, "Macro"),
+            EntryType::Module => write!(f, "Module"),
+            EntryType::Package => write!(f, "Package"),
+            EntryType::Struct => write!(f, "Struct"),
+            EntryType::Trait => write!(f, "Trait"),
+            EntryType::Type => write!(f, "Type")
         }
     }
 }
 
-impl GenerateConfig {
-    fn into_args(self) -> Vec<String> {
-        let mut args = Vec::new();
-        match self.package {
-            Package::Current => {}
-            Package::All => {
-                args.push("--workspace".to_owned());
-            }
-            Package::Single(package_name) => {
-                args.extend_from_slice(&["--package".to_owned(), package_name]);
-            }
-            Package::List(packages) => {
-                for package in packages {
-                    args.extend_from_slice(&["--package".to_owned(), package]);
-                }
-            }
-        }
-        if self.no_dependencies {
-            args.push("--no-deps".to_owned())
-        }
-        if self.doc_private_items {
-            args.push("--document-private-items".to_owned())
-        }
-        if !self.exclude.is_empty() {
-            args.push("--exclude".to_owned());
-            args.extend(self.exclude);
-        }
-        if !self.features.is_empty() {
-            args.push("--features".to_owned());
-            args.extend(self.features);
-        }
-        if self.no_default_features {
-            args.push("--no-default-features".to_owned())
-        }
-        if self.all_features {
-            args.push("--all-features".to_owned())
-        }
-        if let Some(target) = self.target {
-            args.push("--target".to_owned());
-            args.push(target);
-        }
-        if let Some(target_dir) = self.target_dir {
-            args.push("--target-dir".to_owned());
-            args.push(target_dir);
-        }
-        if self.lib {
-            args.push("--lib".to_owned());
-        }
-        if self.bins {
-            args.push("bins".to_owned());
-        }
-        args
-    }
+#[derive(Debug, Clone, PartialEq, Eq, Constructor)]
+pub struct DocsetEntry {
+    pub name: String,
+    pub ty: EntryType,
+    pub path: PathBuf
 }
 
 fn parse_docset_entry<P1: AsRef<Path>, P2: AsRef<Path>>(
@@ -156,37 +97,37 @@ fn parse_docset_entry<P1: AsRef<Path>, P2: AsRef<Path>>(
             }
             3 => match parts[0] {
                 "const" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Constant,
                     file_db_path
                 )),
                 "enum" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Enum,
                     file_db_path
                 )),
                 "fn" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Function,
                     file_db_path
                 )),
                 "macro" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Macro,
                     file_db_path
                 )),
                 "trait" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Trait,
                     file_db_path
                 )),
                 "struct" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Struct,
                     file_db_path
                 )),
                 "type" => Some(DocsetEntry::new(
-                    format!("{}::{}", module_path.unwrap().to_string(), parts[1]),
+                    format!("{}::{}", module_path.unwrap(), parts[1]),
                     EntryType::Type,
                     file_db_path
                 )),
@@ -206,7 +147,7 @@ fn recursive_walk(
     cur_dir: &Path,
     module_path: Option<&str>
 ) -> Result<Vec<DocsetEntry>> {
-    let dir = read_dir(cur_dir).context(IoRead)?;
+    let dir = read_dir(cur_dir).context(IoReadSnafu)?;
     let mut entries = vec![];
     let mut subdir_entries = vec![];
 
@@ -221,7 +162,7 @@ fn recursive_walk(
             if !(module_path.is_none() && ROOT_SKIP_DIRS.contains(&dir_name.as_str())) {
                 subdir_module_path.push_str(&dir_name);
                 subdir_entries.push(recursive_walk(
-                    &root_dir,
+                    root_dir,
                     &dir_entry.path(),
                     Some(&subdir_module_path)
                 ));
@@ -241,36 +182,36 @@ fn generate_sqlite_index<P: AsRef<Path>>(docset_dir: P, entries: Vec<DocsetEntry
     conn_path.push("Contents");
     conn_path.push("Resources");
     conn_path.push("docSet.dsidx");
-    let mut conn = Connection::open(&conn_path).context(Sqlite)?;
+    let mut conn = Connection::open(&conn_path).context(SqliteSnafu)?;
     conn.execute(
         "CREATE TABLE searchIndex(id INTEGER PRIMARY KEY, name TEXT, type TEXT, path TEXT);
         CREATE UNIQUE INDEX anchor ON searchIndex (name, type, path);
         )",
         params![]
     )
-    .context(Sqlite)?;
-    let transaction = conn.transaction().context(Sqlite)?;
+    .context(SqliteSnafu)?;
+    let transaction = conn.transaction().context(SqliteSnafu)?;
     {
         let mut stmt = transaction
             .prepare("INSERT INTO searchIndex (name, type, path) VALUES (?1, ?2, ?3)")
-            .context(Sqlite)?;
+            .context(SqliteSnafu)?;
         for entry in entries {
             stmt.execute([
                 entry.name,
                 entry.ty.to_string(),
                 entry.path.to_str().unwrap().to_owned()
             ])
-            .context(Sqlite)?;
+            .context(SqliteSnafu)?;
         }
     }
-    transaction.commit().context(Sqlite)?;
+    transaction.commit().context(SqliteSnafu)?;
     Ok(())
 }
 
 fn copy_dir_recursive<Ps: AsRef<Path>, Pd: AsRef<Path>>(src: Ps, dst: Pd) -> Result<()> {
-    create_dir_all(&dst).context(IoWrite)?;
-    for entry in read_dir(&src).context(IoRead)? {
-        let entry = entry.context(IoWrite)?.path();
+    create_dir_all(&dst).context(IoWriteSnafu)?;
+    for entry in read_dir(&src).context(IoReadSnafu)? {
+        let entry = entry.context(IoWriteSnafu)?.path();
         if entry.is_dir() {
             let mut dst_dir = dst.as_ref().to_owned();
             dst_dir.push(entry.strip_prefix(&src).unwrap());
@@ -278,147 +219,173 @@ fn copy_dir_recursive<Ps: AsRef<Path>, Pd: AsRef<Path>>(src: Ps, dst: Pd) -> Res
         } else if entry.is_file() {
             let mut dst_file = dst.as_ref().to_owned();
             dst_file.push(entry.file_name().unwrap());
-            copy(entry, dst_file).context(IoWrite)?;
+            copy(entry, dst_file).context(IoWriteSnafu)?;
         }
     }
     Ok(())
 }
 
-fn write_metadata<P: AsRef<Path>>(docset_root_dir: P, package_name: &str, is_virtual_manifest: bool, first_package_name: Option<&str>) -> Result<()> {
+fn write_metadata<P: AsRef<Path>>(
+    docset_root_dir: P,
+    docset_name: &str,
+    index_package: Option<String>,
+    platform_family: Option<String>
+) -> Result<()> {
     let mut info_plist_path = docset_root_dir.as_ref().to_owned();
     info_plist_path.push("Contents");
     info_plist_path.push("Info.plist");
 
-    let mut info_file = File::create(info_plist_path).context(IoWrite)?;
-    let docset_index = if is_virtual_manifest {
-        format!("{}/index.html", first_package_name.unwrap())
+    let mut info_file = File::create(info_plist_path).context(IoWriteSnafu)?;
+    let index_entry = if let Some(index_package) = index_package {
+        format!(
+            "<key>dashIndexFilePath</key>
+                    <string>{}/index.html</string>",
+            index_package
+        )
     } else {
-        format!("{}/index.html", package_name)
+        String::new()
     };
+    let identifier_entry = if let Some(platform_family) = &platform_family {
+        format!(
+            "<key>CFBundleIdentifier</key>
+                    <string>{}</string>",
+            platform_family
+        )
+    } else {
+        String::new()
+    };
+
+    let platform_family_entry = if let Some(platform_family) = &platform_family {
+        format!(
+            "<key>DocSetPlatformFamily</key>
+                    <string>{}</string>",
+            platform_family
+        )
+    } else {
+        String::new()
+    };
+
     write!(info_file,
         "\
         <?xml version=\"1.0\" encoding=\"UTF-8\"?>
         <!DOCTYPE plist PUBLIC \"-//Apple//DTD PLIST 1.0//EN\" \"http://www.apple.com/DTDs/PropertyList-1.0.dtd\">
         <plist version=\"1.0\">
         <dict>
-            <key>CFBundleIdentifier</key>
-                <string>{}</string>
+            {}
             <key>CFBundleName</key>
                 <string>{}</string>
-            <key>dashIndexFilePath</key>
-                <string>{}</string>
-            <key>DocSetPlatformFamily</key>
-                <string>{}</string>
+            {}
+            {}
             <key>isDashDocset</key>
                 <true/>
             <key>isJavaScriptEnabled</key>
                 <true/>
         </dict>
         </plist>",
-         package_name, package_name, docset_index, package_name).context(IoWrite)?;
+        identifier_entry, docset_name, index_entry, platform_family_entry).context(IoWriteSnafu)?;
     Ok(())
 }
 
-pub fn generate(cfg: GenerateConfig) -> Result<()> {
+fn get_workspace_name(metadata: &Metadata) -> String {
+    metadata.workspace_root
+        .file_name()
+        // I doubt this could be None, but let's have a fallback just in case.
+        .unwrap_or("generated-docset")
+        .to_owned()
+}
+
+/// Determine the name we will use for the generated docset.
+/// If a name was provided on the command line, we use this one.
+/// If no name was provided:
+///   * If a single package was requested, use this one.
+///   * Otherwise, if there is a workspace root package and we have been asked to generate
+///     documentation for it, use this one.
+///   * Otherwise, generate a name from the workspace "name" and the list of workspace member packages being built.
+fn get_docset_name(cfg: &DocsetParams, metadata: &Metadata) -> String {
+    if let Some(docset_name) = &cfg.docset_name {
+        return docset_name.to_owned();
+    }
+
+    let (included, _excluded) = cfg.workspace.partition_packages(metadata);
+
+    if included.len() == 1 {
+        return included[0].name.to_owned();
+    }
+
+    // Rust 1.64 will stabilize the `let_chains` feature which should allow combining both conditionals
+    if let Some(root_package) = metadata.root_package() {
+        if included.contains(&root_package) {
+            return root_package.name.to_owned();
+        }
+    }
+
+    let package_list_str = included
+        .into_iter()
+        .filter_map(|p|  {
+            if cfg.workspace.exclude.contains(&p.name) { None } else { Some(p.name.as_str()) }
+        })
+        .collect::<Vec<_>>()
+        .join(", ");
+
+    format!("{}: {}", get_workspace_name(metadata), package_list_str)
+}
+
+/// Return the name of the package that should be used for the docset index, if any.
+/// This uses the same rules as docset name selection, except no index is a valid option.
+fn get_docset_index(cfg: &DocsetParams, metadata: &Metadata) -> Option<String> {
+    if cfg.docset_index.is_some() {
+        return cfg.docset_index.clone();
+    }
+
+    match (cfg.workspace.all, cfg.workspace.package.len()) {
+        (false, 1) => Some(cfg.workspace.package[0].to_owned()),
+        _ => metadata.root_package().map(|p| p.name.to_owned())
+    }
+}
+
+/// Return the keywork that should be used for the docset platform family, if any.
+/// This uses the same rules as docset name selection, except no identifier is a valid option.
+fn get_docset_platform_family(cfg: &DocsetParams, metadata: &Metadata) -> Option<String> {
+    if let Some(platform_family) = &cfg.platform_family {
+        return Some(platform_family.to_owned());
+    }
+
+    match (cfg.workspace.all, cfg.workspace.package.len()) {
+        (false, 1) => Some(cfg.workspace.package[0].to_owned()),
+        _ => metadata.root_package().map(|p| p.name.to_owned())
+    }
+}
+
+pub fn generate_docset(cfg: DocsetParams) -> Result<()> {
     // Step 1: generate rustdoc
     // Figure out for which crate to build the doc and invoke cargo doc.
     // If no crate is specified, run cargo doc for the current crate/workspace.
-    if cfg.package != Package::All && !cfg.exclude.is_empty() {
-        return Args {
-            msg: "--exclude must be used with --all"
-        }
-        .fail();
-    }
-    // Determine the package name that we will use for the generated docset.
-    // If a single package was requested, use this one.
-    // Otherwise, we use the "root" package/workspace name.
-    // TODO: we should probably handle the Package::List case differently. Maybe provide an option
-    // to set the generated docset name ?
-    let mut is_virtual_manifest = false;
-    let mut first_package_name = None;
-    let package_name = match cfg.package.clone() {
-        Package::Single(package) => package,
-        _ => {
-            let manifest_location =
-                cfg.manifest_path
-                    .clone()
-                    .unwrap_or(locate_package_manifest()?);
-            let mut manifest_file = File::open(&manifest_location)
-                .context(IoRead)?;
-            let mut manifest_contents = String::new();
-            manifest_file
-                .read_to_string(&mut manifest_contents)
-                .context(IoRead)?;
-            let toml_manifest = manifest_contents.parse::<Value>().context(Toml)?;
-            let package_table = toml_manifest
-                .as_table()
-                .expect("Cargo.toml is not a toml table")
-                .get("package");
-            match package_table {
-                Some(toml) => {
-                    toml.as_table()
-                    .expect("Cargo.toml package entry is not a toml table")
-                    .get("name")
-                    .expect("Cargo.toml doesn't define a package name")
-                    .as_str()
-                    .expect("Cargo.toml package.name entry is not a string")
-                    .to_owned()
-                }
-                None => {
-                    is_virtual_manifest = true;
-                    let members = toml_manifest.as_table()
-                        .unwrap()
-                        .get("workspace")
-                        .expect("Manifest has neither a package nor a workspace section")
-                        .as_table()
-                        .expect("Workspace section is not a toml table")
-                        .get("members")
-                        .expect("Workspace section does not have a member field")
-                        .as_array()
-                        .expect("Members field is not an array");
-                    first_package_name = Some(
-                        members
-                            .first()
-                            .expect("Virtual manifest workspace has no members")
-                            .as_str()
-                            .expect("Workspace member is not a string")
-                            .to_owned()
-                    );
-                    let manifest_path = Path::new(&manifest_location);
-                    manifest_path
-                        .parent()
-                        .expect("Manifest parent location is not a directory ???")
-                        .file_name()
-                        .unwrap()
-                        .to_string_lossy()
-                        .to_string()
-                }
+    if cfg.workspace.all {
+        ensure!(
+            cfg.workspace.exclude.is_empty(),
+            ArgsSnafu {
+                msg: "--exclude must be used with --all"
             }
-        }
-    };
-    let cargo_metadata = get_cargo_metadata()?;
-    let docset_root_dir = cfg.target_dir.as_deref().unwrap_or(&cargo_metadata.target_directory);
-    let mut docset_root_dir_path = PathBuf::from_str(docset_root_dir).unwrap();
-    let mut rustdoc_root_dir = docset_root_dir_path.clone();
-    rustdoc_root_dir.push("doc");
-    docset_root_dir_path.push("docset");
-    docset_root_dir_path.push(format!("{}.docset", package_name));
+        );
+    }
 
-    // Clean the documentation directory if so requested
-    if cfg.clean {
+    let cargo_metadata = cfg.manifest.metadata().exec().context(CargoMetadataSnafu)?;
+
+    // Clean the documentation directory if the user didn't explicitly ask not to clean it.
+    if !cfg.no_clean {
         println!("Running 'cargo clean --doc'...");
         let mut cargo_clean_args = vec!["clean".to_owned()];
-        if let Some(ref manifest_path) = &cfg.manifest_path {
+        if let Some(ref manifest_path) = &cfg.manifest.manifest_path {
             cargo_clean_args.push("--manifest-path".to_owned());
-            cargo_clean_args.push(manifest_path.to_owned());
+            cargo_clean_args.push(manifest_path.to_string_lossy().to_string());
         }
         let cargo_clean_result = Command::new("cargo")
             .args(cargo_clean_args)
             .arg("--doc")
             .status()
-            .context(Spawn)?;
+            .context(SpawnSnafu)?;
         if !cargo_clean_result.success() {
-            return CargoClean {
+            return CargoCleanSnafu {
                 code: cargo_clean_result.code()
             }
             .fail();
@@ -426,41 +393,67 @@ pub fn generate(cfg: GenerateConfig) -> Result<()> {
     }
     // Good to go, generate the documentation.
     println!("Running 'cargo doc'...");
+    let args = cfg.clone().into_args();
     let cargo_doc_result = Command::new("cargo")
         .arg("doc")
-        .args(cfg.into_args())
+        .args(args)
         .status()
-        .context(Spawn)?;
+        .context(SpawnSnafu)?;
     if !cargo_doc_result.success() {
-        return CargoDoc {
+        return CargoDocSnafu {
             code: cargo_doc_result.code()
         }
         .fail();
     }
 
     // Step 2: iterate over all the html files in the doc directory and parse the filenames
+    let docset_name = get_docset_name(&cfg, &cargo_metadata);
+    let mut docset_root_dir = cfg
+        .target_dir
+        .clone()
+        .unwrap_or_else(|| cargo_metadata.target_directory.clone().into_std_path_buf());
+    let mut rustdoc_root_dir = docset_root_dir.clone();
+    rustdoc_root_dir.push("doc");
+    docset_root_dir.push("docset");
+    let docset_identifier = get_docset_platform_family(&cfg, &cargo_metadata);
+    docset_root_dir.push(
+        format!("{}.docset",
+            docset_identifier.clone()
+                .unwrap_or_else(|| get_workspace_name(&cargo_metadata))));
     let entries = recursive_walk(&rustdoc_root_dir, &rustdoc_root_dir, None)?;
 
     // Step 3: generate the SQLite database
     // At this point, we need to start writing into the output docset directory, so create the
     // hirerarchy, and clean it first if it already exists.
-    if docset_root_dir_path.exists() {
-        remove_dir_all(&docset_root_dir_path).context(IoWrite)?;
+    if docset_root_dir.exists() {
+        remove_dir_all(&docset_root_dir).context(IoWriteSnafu)?;
     }
-    let mut docset_hierarchy = docset_root_dir_path.clone();
+    let mut docset_hierarchy = docset_root_dir.clone();
     docset_hierarchy.push("Contents");
     docset_hierarchy.push("Resources");
-    create_dir_all(&docset_hierarchy).context(IoWrite)?;
-    generate_sqlite_index(&docset_root_dir_path, entries)?;
+    create_dir_all(&docset_hierarchy).context(IoWriteSnafu)?;
+    generate_sqlite_index(&docset_root_dir, entries)?;
 
     // Step 4: Copy the rustdoc to the docset directory
     docset_hierarchy.push("Documents");
     copy_dir_recursive(&rustdoc_root_dir, &docset_hierarchy)?;
 
     // Step 5: add the required metadata
-    write_metadata(&docset_root_dir_path, &package_name, is_virtual_manifest, first_package_name.as_deref())?;
+    if docset_identifier.is_none() {
+        warn("no docset identifier was provided and none could be generated, consider adding the '--docset-identifier' option.");
+    }
 
-    println!("Docset succesfully generated in {}", docset_root_dir_path.to_string_lossy());
+    write_metadata(
+        &docset_root_dir,
+        &docset_name,
+        get_docset_index(&cfg, &cargo_metadata),
+        docset_identifier
+    )?;
+
+    println!(
+        "Docset succesfully generated in {}",
+        docset_root_dir.to_string_lossy()
+    );
 
     Ok(())
 }