+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # cargo-docset changelog
 
+## Unreleased - v0.2.2
+
+* Bugfix: module names are no longer suffixed by `::index`.
+* Bugfix: fix several issues with virtual workspaces.
+* Feature: add `--target-dir` option and respect `CARGO_TARGET_DIR` environment variable and `build.target_dir` config.
+* Feature: add the `--docset-name` option in order to specify or override the docset name.
+* Feature: add the `--docset-index` option in order to specify or override the docset index package.
+* Feature: add the `--platform-family` option in order to specify or override the docset platform family string.
+* Feature: add the enabled by default `color` cargo feature which can be disabled to turn off colored terminal output.
+* Refactored: use the [cargo-metadata](https://crates.io/crates/cargo_metadata) crate to obtain the workspace metadata,
+  replace hand-rolled mechanisms.
+* Maintenance: update dependencies to their latest versions.
+* Maintenance: update to Rust edition 2021.
+
 ## 8/23/2020 - v0.2.1
 
 * Bugfix: fix spelling of the `manifest-path` when passed down to cargo.